# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Parcel tracking service",
        "title": "Delivery Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/parcels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список посылок с пагинацией и фильтром по статусу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Получение списка посылок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parcel status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.ListParcelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт посылку и генерирует трек-номер",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Создание посылки",
                "parameters": [
                    {
                        "description": "Create parcel request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.CreateParcelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.CreateParcelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parcels/archive": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отправляет посылку в архив",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Архивирует посылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parcels/photos": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает фото посылки и сохраняет путь к файлу",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Добавление фото к посылке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parcel ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Photo file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parcels/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет статус посылки и добавляет запись в историю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Обновление статуса посылки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Change status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.ChangeStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parcels/track": {
            "get": {
                "description": "Возвращает информацию о посылке, историю статусов и фотографии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parcel"
                ],
                "summary": "Получение посылки по трек-номеру",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track number",
                        "name": "track_number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.GetParcelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "delivery-tracker_internal_domain.Status": {
            "type": "string",
            "enum": [
                "CREATED",
                "PURCHASED",
                "WAREHOUSE",
                "IN_TRANSIT",
                "CUSTOMS",
                "ARRIVED",
                "DELIVERED"
            ],
            "x-enum-varnames": [
                "StatusCreated",
                "StatusPurchased",
                "StatusWarehouse",
                "StatusInTransit",
                "StatusCustoms",
                "StatusArrived",
                "StatusDelivered"
            ]
        },
        "delivery-tracker_internal_dto.ChangeStatusRequest": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/delivery-tracker_internal_domain.Status"
                }
            }
        },
        "delivery-tracker_internal_dto.CreateParcelRequest": {
            "type": "object",
            "properties": {
                "item_name": {
                    "type": "string"
                },
                "recipient_address": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "recipient_phone": {
                    "type": "string"
                }
            }
        },
        "delivery-tracker_internal_dto.CreateParcelResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
...
        "delivery-tracker_internal_dto.GetParcelResponse": {
            "type": "object",
            "properties": {
                "current_location": {
                    "type": "string"
                },
                "current_status": {
                    "$ref": "#/definitions/delivery-tracker_internal_domain.Status"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery-tracker_internal_dto.ParcelHistoryResponse"
                    }
                },
                "item_name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery-tracker_internal_dto.ParcelPhotoResponse"
                    }
                },
                "recipient": {
                    "type": "string"
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "delivery-tracker_internal_dto.ListParcelResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery-tracker_internal_dto.ParcelListItemResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "delivery-tracker_internal_dto.ParcelHistoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "new_status": {
                    "$ref": "#/definitions/delivery-tracker_internal_domain.Status"
                },
                "old_status": {
                    "$ref": "#/definitions/delivery-tracker_internal_domain.Status"
                }
            }
        },
        "delivery-tracker_internal_dto.ParcelListItemResponse": {
            "type": "object",
            "properties": {
                "current_location": {
                    "type": "string"
                },
                "current_status": {
                    "$ref": "#/definitions/delivery-tracker_internal_domain.Status"
                },
                "id": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "delivery-tracker_internal_dto.ParcelPhotoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                }
            }
        },
        "delivery-tracker_internal_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}